requirements.txt
gunicorn==20.1.0
Django==4.0.2
djangorestframework==3.13.1
django-filter==21.1
django-parler==2.3
django-ckeditor==6.1.0
psycopg2-binary==2.9.3
beautifulsoup4==4.10.0
Pillow==9.0.0
lxml==4.8.0
